Scaffold of CRDs needed for kube edge
Basic scaffolding of a operator with operator-sdk.
Created two CRDs: edgenode and edgedeployment - should we follow best
practices and create separate operators?

Need to think about what additional CRDs we want to create? What will be
the best way for the "agent" to parse an object to find a valid PodSpec?